refactor: Group all ROOT writers together
