+++ Examples/Io/Root/src/TrackFinderPerformanceWriter.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "ActsExamples/Io/Performance/TrackFinderPerformanceWriter.hpp"
+#include "ActsExamples/Io/Root/TrackFinderPerformanceWriter.hpp"
 
 #include "Acts/Definitions/Units.hpp"
 #include "ActsExamples/EventData/Index.hpp"
