+++ Examples/Io/Root/src/detail/NuclearInteractionParametrisation.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "ActsExamples/Io/NuclearInteractions/detail/NuclearInteractionParametrisation.hpp"
+#include "ActsExamples/Io/Root/detail/NuclearInteractionParametrisation.hpp"
 
 #include "Acts/Definitions/Common.hpp"
 #include "ActsFatras/EventData/Particle.hpp"
